moved all the scripts from models to ml_pipline
