+++ backend/ml_pipeline/RAG/setup_embeddings.py
@@ -1,4 +1,4 @@
-from backend.models.RAG.src.embeddings import EmbeddingsHandler
+from backend.ml_pipeline.RAG.embeddings import EmbeddingsHandler
 print("working")
 def setup():
     print("Starting one-time embedding generation...")
